drivers: pca9548: move calibration to new 'shell_menu'
Calibration does not exist anymore.
Move shell commands for the pca9548 into a dedicated module
'shell_pca9548'.
