--- drivers/include/calibration/calib_pca9548.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "pca9548.h"
-
-void pca9548_calib_init(void);
