+++ drivers/pca9548/Makefile
@@ -1,3 +1,1 @@
-MODULE = pca9548
-
 include $(RIOTBASE)/Makefile.base
