+++ drivers/pca9548/shell_pca9548/Makefile
@@ -1,3 +1,1 @@
-MODULE = calib_pca9548
-
 include $(RIOTBASE)/Makefile.base
