+++ drivers/pca9548/shell_pca9548/Makefile.dep
@@ -1,0 +1,2 @@
+USEMODULE += pca9548
+USEMODULE += shell_menu
