+++ drivers/pca9548/shell_pca9548/include/shell_pca9548.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void pca9548_shell_init(void);
